wip: cleanup, fixes, helper functions
apologies for the commit mess, i was bored at home and just walked
through the code and started cleaning up and fixing stuff.

mostly structuring errors into something reusable and easy to work with.
creating functions to help with that, e.g., parseDBError and
parseValidationErrors.

i know it's more boilerplate but it makes it nice to work with.